"Use useSelector hook in App.js and EnrolledCourses.jsx to access user data from Redux store"
This commit updates the files App.js and EnrolledCourses.jsx to use the useSelector hook from the 'react-redux' library to access user data from the Redux store. In App.js, the user data is used to conditionally render certain routes based on the user's account type. In EnrolledCourses.jsx, the user's token is used to fetch the user's enrolled courses from the server. The token is obtained from the Redux store using the useSelector hook. Additionally, the commit updates the profileSlice.js file to move the user data to the top level of the state and add a loading field.